requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Deprecated==1.2.12
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
opentelemetry-api==1.3.0
opentelemetry-instrumentation==0.22b0
opentelemetry-instrumentation-flask==0.22b0
opentelemetry-instrumentation-requests==0.22b0
opentelemetry-instrumentation-wsgi==0.22b0
opentelemetry-sdk==1.3.0
opentelemetry-semantic-conventions==0.22b0
opentelemetry-util-http==0.22b0
requests==2.25.1
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
